+++ src/main/java/smd/morebaubles/entity/EntityTerrariaThrowable.java
@@ -127,9 +127,9 @@ public abstract class EntityTerrariaThrowable extends Entity implements IProject
 			}
 
 			this.inGround = false;
-			this.motionX *= (double) (this.rand.nextFloat()*0.2F);
-			this.motionY *= (double) (this.rand.nextFloat()*0.2F);
-			this.motionZ *= (double) (this.rand.nextFloat()*0.2F);
+			this.motionX *= this.rand.nextFloat()*0.2F;
+			this.motionY *= this.rand.nextFloat()*0.2F;
+			this.motionZ *= this.rand.nextFloat()*0.2F;
 			this.ticksInGround = 0;
 			this.ticksInAir = 0;
 		} else {
@@ -209,7 +209,7 @@ public abstract class EntityTerrariaThrowable extends Entity implements IProject
 		this.posZ += this.motionZ;
 		float f = MathHelper.sqrt(this.motionX*this.motionX+this.motionZ*this.motionZ);
 		this.rotationYaw = (float) (MathHelper.atan2(this.motionX, this.motionZ)*(180D/Math.PI));
-		this.rotationPitch = (float) (MathHelper.atan2(this.motionY, (double) f)*(180D/Math.PI));
+		this.rotationPitch = (float) (MathHelper.atan2(this.motionY, f)*(180D/Math.PI));
 		while ((this.rotationPitch = this.prevRotationPitch)<-180.0F) {
 			this.prevRotationPitch -= 360.0F;
 		}
@@ -242,12 +242,12 @@ public abstract class EntityTerrariaThrowable extends Entity implements IProject
 			f1 = 0.8F;
 		}
 
-		this.motionX *= (double) f1;
-		this.motionY *= (double) f1;
-		this.motionZ *= (double) f1;
+		this.motionX *= f1;
+		this.motionY *= f1;
+		this.motionZ *= f1;
 
 		if (!this.hasNoGravity()) {
-			this.motionY -= (double) f2;
+			this.motionY -= f2;
 		}
 
 		this.setPosition(this.posX, this.posY, this.posZ);
@@ -333,7 +333,7 @@ public abstract class EntityTerrariaThrowable extends Entity implements IProject
 		float f1 = -MathHelper.sin((rotationPitchIn+pitchOffset)*0.017453292F);
 		float f2 = MathHelper.cos(rotationYawIn*0.017453292F)
 				*MathHelper.cos(rotationPitchIn*0.017453292F);
-		this.shoot((double) f, (double) f1, (double) f2, velocity, inaccuracy);
+		this.shoot(f, f1, f2, velocity, inaccuracy);
 		this.motionX += entityThrower.motionX;
 		this.motionZ += entityThrower.motionZ;
 
@@ -359,7 +359,7 @@ public abstract class EntityTerrariaThrowable extends Entity implements IProject
 		this.motionZ = z;
 		float f1 = MathHelper.sqrt(x*x+z*z);
 		this.rotationYaw = (float) (MathHelper.atan2(x, z)*(180D/Math.PI));
-		this.rotationPitch = (float) (MathHelper.atan2(y, (double) f1)*(180D/Math.PI));
+		this.rotationPitch = (float) (MathHelper.atan2(y, f1)*(180D/Math.PI));
 		this.prevRotationYaw = this.rotationYaw;
 		this.prevRotationPitch = this.rotationPitch;
 		this.ticksInGround = 0;
